rename eval module
